Address CodeRabbit review: isolate zip extraction from Django imports.
Move safe_extract_zip_file into base.zip_utils for workers without Django
setup and close zip handles on extraction failure.
