+++ scripts/workers/remote_submission_worker.py
@@ -134,7 +134,7 @@ def download_and_extract_zip_file(url, download_location, extract_location):
         with open(download_location, "wb") as f:
             f.write(response.content)
         # extract zip file
-        from base.utils import safe_extract_zip_file
+        from base.zip_utils import safe_extract_zip_file
 
         zip_ref = zipfile.ZipFile(download_location, "r")
         safe_extract_zip_file(zip_ref, extract_location)
