+++ scripts/workers/submission_worker.py
@@ -177,7 +177,7 @@ def extract_zip_file(download_location, extract_location):
         * `download_location`: Location of zip file
         * `extract_location`: Location of directory for extracted file
     """
-    from base.utils import safe_extract_zip_file
+    from base.zip_utils import safe_extract_zip_file
 
     zip_ref = zipfile.ZipFile(download_location, "r")
     safe_extract_zip_file(zip_ref, extract_location)
